service_users/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
python-dotenv
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]
redis
pydantic[email]>=2
pydantic-settings>=2.2
sendgrid>=6.11.0